+++ pkg/api/types.go
@@ -4172,7 +4172,7 @@ type SecurityContext struct {
 	// +optional
 	ReadOnlyRootFilesystem *bool
 	// AllowPrivilegeEscalation controls whether a process can gain more
-	// privileges than it's parent process. This bool directly controls if
+	// privileges than its parent process. This bool directly controls if
 	// the no_new_privs flag will be set on the container process.
 	// +optional
 	AllowPrivilegeEscalation *bool