+++ plugin/pkg/scheduler/algorithm/predicates/metadata.go
@@ -18,13 +18,15 @@ package predicates
 
 import (
 	"fmt"
-	"github.com/golang/glog"
+	"sync"
+
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/kubernetes/plugin/pkg/scheduler/algorithm"
 	"k8s.io/kubernetes/plugin/pkg/scheduler/schedulercache"
 	schedutil "k8s.io/kubernetes/plugin/pkg/scheduler/util"
-	"sync"
+
+	"github.com/golang/glog"
 )
 
 type PredicateMetadataFactory struct {