+++ plugin/pkg/scheduler/schedulercache/cache.go
@@ -21,10 +21,11 @@ import (
 	"sync"
 	"time"
 
-	"github.com/golang/glog"
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/util/wait"
+
+	"github.com/golang/glog"
 )
 
 var (